requirements.txt
Flask==3.1.2
flask-bcrypt==1.0.1
pymongo==4.15.2
python-dotenv==1.2.1
certifi>=2024.2.2    